+++ R/model_get_model_matrix.R
@@ -130,3 +130,12 @@ model_get_model_matrix.mmrm <- function(model, ...) {
     data = stats::model.frame(model)
   )
 }
+
+#' @export
+#' @rdname model_get_model_matrix
+model_get_model_matrix.betareg <- function(model, ...) {
+  stats::model.matrix.default(
+    model %>% model_get_terms(),
+    data = model %>% model_get_model_frame()
+  )
+}