+++ data-raw/DATASET.R
@@ -28,7 +28,7 @@ supported_models <-
     "`MASS::glm.nb()`", "",
     "`mice::mira`", paste(
       "Limited support. If `mod` is a `mira` object, use",
-      "`tidy_plus_plus(mod, tidy_fun = function(x, ...) mice::pool(x) %>% mice::tidy(...))`"
+      "`tidy_fun = function(x, ...) {mice::pool(x) %>% mice::tidy(...)}`"
     ),
     "`lavaan::lavaan()`", "Limited support for categorical variables",
     "`stats::nls()`", "Limited support",
@@ -51,6 +51,9 @@ supported_models <-
     "`multgee::nomLORgee()`", "Experimental support. Use `tidy_multgee()` as `tidy_fun`.",
     "`multgee::ordLORgee()`", "Experimental support. Use `tidy_multgee()` as `tidy_fun`.",
     "`mmrm::mmrm()`", "",
+    "`pscl::zeroinfl()`", "Use `tidy_zeroinfl()` as `tidy_fun`.",
+    "`pscl::hurdle()`", "Use `tidy_zeroinfl()` as `tidy_fun`.",
+    "`betareg::betareg()`", "Use `tidy_parameters()` as `tidy_fun` with `component` argument to control with coefficients to return. `broom::tidy()` does not support the `exponentiate` argument for betareg models, use `tidy_parameters()` instead." # nolint
   ) %>%
   dplyr::arrange(.data$model, .locale = "en")